Change packages to internal :)
We don't really need `pkg/catalog-cd` as a package, but we do want to,
at least for now, prevent anyone outside `openshift-pipelines/tektoncd-catalog`
to import code as it's only purpose is to be used from here.

This might change in the future, but we can then extract/export code
as needed.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>